delete images now prevents deletion of other user's images

+++ models/images-model.js
@@ -42,6 +42,9 @@ async function deleteImages(userId, imageIds){
     let images = await db("images").whereIn("id", imageIds);
     
     for (let image of images){
+        if (userId !== image.user_id){
+            throw new Error("that image does not belong to this user")
+        }
         await fsPromises.unlink(image.file_path);
     }
